refactor(architecture): restore source size boundaries
