+++ src/features/internal-storage/main/infrastructure/worker/internalStorageMigrations.ts
@@ -8,24 +8,12 @@ import {
 
 import type DatabaseConstructor from 'better-sqlite3';
 
+export { INTERNAL_STORAGE_REQUIRED_BACKUP_TABLES } from './internalStorageBackupTables';
+
 type SqliteDatabase = InstanceType<typeof DatabaseConstructor>;
 
 /** "ATAI" in big-endian ASCII. Backups reject databases owned by another application. */
 export const INTERNAL_STORAGE_APPLICATION_ID = 0x41544149;
-export const INTERNAL_STORAGE_REQUIRED_BACKUP_TABLES = Object.freeze([
-  'coordination_backup_runs',
-  'coordination_backup_writer_fences',
-  'coordination_event_journal',
-  'coordination_event_journal_metadata',
-  'durable_application_command_outbox',
-  'durable_application_commands',
-  'snapshot_retention_leases',
-  'team_identity_records',
-  'team_identity_storage_metadata',
-  'team_roster_members',
-  'team_roster_storage_metadata',
-  'team_rosters',
-] as const);
 
 interface InternalStorageMigration {
   version: number;
